# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  models.Employment:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Exhibitor:
    properties:
      about:
        type: string
      cities:
        type: string
      climateCompensation:
        type: boolean
      companyWebsite:
        type: string
      employments:
        items:
          $ref: '#/definitions/models.Employment'
        type: array
      eventroId:
        type: string
      fairLocation:
        type: string
      flyer:
        type: string
      id:
        type: integer
      industries:
        items:
          $ref: '#/definitions/models.Industry'
        type: array
      locationSpecial:
        type: string
      logoFreesize:
        type: string
      logoSquared:
        type: string
      mapImg:
        type: string
      name:
        type: string
      programs:
        items:
          $ref: '#/definitions/models.Program'
        type: array
      purpose:
        type: string
      tier:
        $ref: '#/definitions/models.Tier'
      type:
        type: string
      vyerPosition:
        type: string
    type: object
...
  models.Industry:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Program:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
  models.Tier:
    enum:
    - Standard
    - Bronze
    - Silver
    - Gold
    type: string
    x-enum-varnames:
    - TierStandard
    - TierBronze
    - TierSilver
    - TierGold
...
info:
  contact:
    name: THS Armada Development
    url: https://github.com/armada-ths
  description: REST API powering the THS Armada career fair website (armada.nu) and
    its admin panel.
  title: ArmadaCMS API
  version: "1.0"
paths:
...
  /exhibitors:
    get:
      parameters:
      - description: Pagination range, e.g. [0,24]
        in: query
        name: range
        type: string
      - description: Sort, e.g. [\
        in: query
        name: sort
        type: string
      - description: Filter, e.g. {\
        in: query
        name: filter
        type: string
      - description: Return all exhibitors without pagination, sorted by tier
        in: query
        name: all
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            Content-Range:
              description: exhibitors 0-24/100
              type: string
          schema:
            items:
              $ref: '#/definitions/models.Exhibitor'
            type: array
      summary: List exhibitors
      tags:
      - exhibitors
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Company name
        in: formData
        name: name
        required: true
        type: string
      - description: Company type
        in: formData
        name: type
        required: true
        type: string
      - description: Fair location
        in: formData
        name: fairLocation
        required: true
        type: string
      - description: Tier (Standard/Bronze/Silver/Gold)
        in: formData
        name: tier
        type: string
      - description: About the company
        in: formData
        name: about
        type: string
      - description: Purpose/mission
        in: formData
        name: purpose
        type: string
      - description: Company website URL
        in: formData
        name: companyWebsite
        type: string
      - description: Climate compensation
        in: formData
        name: climateCompensation
        type: boolean
      - description: City/cities
        in: formData
        name: cities
        type: string
      - description: Logo image (JPG/PNG/WEBP/GIF)
        in: formData
        name: file
        type: file
      - description: JSON array of program IDs, e.g. [1,2,3]
        in: formData
        name: programs
        type: string
      - description: JSON array of industry IDs
        in: formData
        name: industries
        type: string
      - description: JSON array of employment IDs
        in: formData
        name: employments
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.Exhibitor'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Create failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Create exhibitor
      tags:
      - exhibitors
  /exhibitors/{id}:
    delete:
      parameters:
      - description: Exhibitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Deleted
          schema:
            type: string
        "404":
          description: Exhibitor not found
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Delete exhibitor
      tags:
      - exhibitors
    get:
      parameters:
      - description: Exhibitor ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exhibitor'
        "404":
          description: Exhibitor not found
          schema:
            type: string
      summary: Get exhibitor by ID
      tags:
      - exhibitors
    put:
      consumes:
      - multipart/form-data
      parameters:
      - description: Exhibitor ID
        in: path
        name: id
        required: true
        type: integer
      - description: Company name
        in: formData
        name: name
        type: string
      - description: Company type
        in: formData
        name: type
        type: string
      - description: Tier (Standard/Bronze/Silver/Gold)
        in: formData
        name: tier
        type: string
      - description: Logo image (JPG/PNG/WEBP/GIF)
        in: formData
        name: file
        type: file
      - description: JSON array of program IDs
        in: formData
        name: programs
        type: string
      - description: JSON array of industry IDs
        in: formData
        name: industries
        type: string
      - description: JSON array of employment IDs
        in: formData
        name: employments
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Exhibitor'
        "400":
          description: Bad request
          schema:
            type: string
        "404":
          description: Not found
          schema:
            type: string
        "500":
          description: Update failed
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Update exhibitor
      tags:
      - exhibitors
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer JWT token obtained from POST /api/v1/login. Format: "Bearer
      <token>"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"